Minor tweak to how crowd_sim structures are serialized in YAML
By default, it appears that yaml-cpp is always placing empty
sequences on their own lines if the parent map is in block style.
This is different from what the YAML library we're using in
JavaScript is doing, which causes diffs when files are saved
due to just this (meaningless) carriage reutrn. This commit
tells yaml-cpp to use flow style when the sequences are empty
in a few places in the crowd_sim data structure that are,
by default, empty sequences.

Signed-off-by: Morgan Quigley <morgan@osrfoundation.org>